+++ examples/run_examples.jl
@@ -1,10 +1,10 @@
 import QuantumCircuitOpt as QCO
 using JuMP
-using CPLEX
 using Gurobi
-using Cbc
+# using Cbc
+# using CPLEX
 
-include("solver.jl")
+include("optimizers.jl")
 include("2qubit_gates.jl")
 include("3qubit_gates.jl")
 include("4qubit_gates.jl")
@@ -23,11 +23,12 @@ decompose_gates = ["decompose_hadamard",
                    "decompose_swap",
                    "decompose_W",
                    "decompose_W_using_HCnot",
-                   "decompose_GroverDiffusionGate",
+                   "decompose_GroverDiffusion_using_Clifford",
                    "decompose_iSwapGate",
+                   "decompose_qft2_using_HT",
                    "decompose_RX_on_q3"]
 
-# decompose_gates = ["decompose_controlled_Z"]
+decompose_gates = ["decompose_GroverDiffusion_using_HX"]
 
 #----------------------------------------------#
 #      Quantum Circuit Optimization model      #
@@ -35,16 +36,12 @@ decompose_gates = ["decompose_hadamard",
 result_qc = Dict{String,Any}()
 
 for gates in decompose_gates 
+    
     params = getfield(Main, Symbol(gates))()
-    
-    qcm_optimizer = get_solver(params)
-    # qcm_optimizer = JuMP.optimizer_with_attributes(Cbc.Optimizer)
-    # qcm_optimizer = JuMP.optimizer_with_attributes(CPLEX.Optimizer)
-    # qcm_optimizer = JuMP.optimizer_with_attributes(Gurobi.Optimizer, "Presolve" => 1)
-    
-    global result_qc = QCO.run_QCModel(params, 
-                                       qcm_optimizer, 
-                                       model_type = "compact_formulation",
-                                       convex_hull_gate_constraints = false)
-end
 
+    model_options = Dict{Symbol, Any}(:model_type => "compact_formulation",
+                                      :convex_hull_gate_constraints => false)
+    
+    qcopt_optimizer = get_gurobi()
+    global result_qc = QCO.run_QCModel(params, qcopt_optimizer)
+end